Moved item/location/qty cells into own xib files
This way they are not duplicated in Main.Storyboard, and only created in a single xib each

PickerTableViewCell.xib is loaded programatically inside PickerTableViewCell, since it is used by its subclasses. As a result it isn't assigned to any class, and outlets are also assigned programatically.
QtySelectionTableViewCell.xib is only used by its controller SelectQtyTableViewCell, so it is assigned to that class within the xib, and outlets are linked.